pyproject.toml
[project]
name = "alkeymino"
version = "0.1.0"
description = "Generate TCG playing card images from HTML templates."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "playwright>=1.40",
    "pyyaml>=6.0",
]

[project.scripts]
alkeymino = "alkeymino.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alkeymino"]
